+++ spacepixfile.h
@@ -1,7 +1,9 @@
 #pragma once
 
-#include "genlib/p2dpoly.h"
 #include "salalib/shapemap.h"
+
+#include "genlib/p2dpoly.h"
+
 #include <deque>
 #include <string>
 
@@ -13,7 +15,7 @@ class SpacePixelFile {
   public:
     std::deque<ShapeMap> m_spacePixels;
     QtRegion m_region; // easier public for now
-    //
+
     SpacePixelFile(const std::string &name = std::string()) {
         m_name = name;
         m_current_layer = -1;
@@ -33,14 +35,16 @@ class SpacePixelFile {
 
     void setName(const std::string &name) { m_name = name; }
     const std::string &getName() const { return m_name; }
-    //
+
     QtRegion &getRegion() const { return (QtRegion &)m_region; }
-    //
+
     // Screen functionality:
     void makeViewportShapes(const QtRegion &viewport = QtRegion()) const;
     bool findNextShape(bool &nextlayer) const;
 
-    const SalaShape &getNextShape() const { return m_spacePixels[m_current_layer].getNextShape(); }
+    const SalaShape &getNextShape() const {
+        return m_spacePixels[static_cast<size_t>(m_current_layer)].getNextShape();
+    }
 
     // Is any one sublayer shown?
 
@@ -50,7 +54,7 @@ class SpacePixelFile {
                 return true;
         return false;
     }
-    //
+
   public:
     bool read(std::istream &stream);
     bool write(std::ofstream &stream);
